continues implementing custom font component
