+++ src/content/homepage/index.mdoc
@@ -13,7 +13,7 @@ customNavigation:
    bgType={discriminant: "image", value: "/pink-clouds.jpg"} %}
 # Welcome to My Website
 
-Here's {% CustomFont fontSettings={fontFamily: {discriminant: false}, fontWeight: {discriminant: true, value: "900"}, fontStyle: {discriminant: false}, fontSize: {discriminant: true, value: "24px"}, lineHeight: {discriminant: false}, textColor: {discriminant: true, value: "#fa0000"}} %}some{% /CustomFont %} content
+Here's {% CustomFont fontSettings={fontFamily: {discriminant: false}, fontWeight: {discriminant: false}, fontStyle: {discriminant: false}, fontSize: {discriminant: false}, lineHeight: {discriminant: false}, textColor: {discriminant: true, value: "#ad2525"}} customClass="test-class" %}some{% /CustomFont %} content
 {% /PageSection %}
 
 {% PortfolioView
