+++ src/content/pages/about.mdoc
@@ -19,9 +19,9 @@ Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse suscipit lo
 
 [Link to hidden page](/hidden-page)
 
-Aenean congue neque at metus mattis, vel porttitor ante feugiat. Vestibulum urna nulla, mollis vel tristique nec, porttitor vitae lorem. Curabitur suscipit ante tortor, sed ultricies urna suscipit tempus. Donec sed feugiat lorem. Fusce leo mauris, vehicula at risus id, aliquet scelerisque ex. Etiam dignissim ante dolor, non tristique justo auctor sed. Donec magna mauris, bibendum ac dignissim iaculis, mattis vitae tellus. Sed sagittis purus ac massa suscipit sollicitudin eget ac nibh. Fusce sodales tempor ligula, quis accumsan elit viverra sit amet. Sed id porta libero. Suspendisse potenti. Sed lectus sem, commodo et volutpat non, aliquam et leo. Suspendisse pharetra leo sed eros pellentesque finibus. In eget urna at elit tempus ultrices in cursus massa.
+Aenean congue neque at metus mattis, vel porttitor ante feugiat. Vestibulum urna nulla, mollis vel tristique nec, porttitor vitae lorem. Curabitur suscipit ante tortor, sed ultricies urna suscipit tempus. Donec sed feugiat lorem. Fusce leo mauris, vehicula at risus id, aliquet scelerisque ex. Etiam dignissim ante dolor, non tristique justo auctor sed. Donec magna mauris, bibendum ac dignissim iaculis, mattis vitae tellus. Sed sagittis purus ac massa suscipit sollicitudin eget ac nibh. Fusce sodales te{% CustomFont fontSettings={fontFamily: {discriminant: false}, fontWeight: {discriminant: false}, fontStyle: {discriminant: false}, fontSize: {discriminant: true, value: "24px"}, lineHeight: {discriminant: false}, textColor: {discriminant: false}} %}mpor ligula, quis accumsan elit viverra sit amet. Sed id porta libero. Suspendisse potenti. Sed lectus sem, commodo et volutpat non, aliquam et leo. Suspendisse pharetra leo sed eros pellentesque finibus. In eget urna at elit tempus ultrices in cursus massa.{% /CustomFont %}
 
-Mauris gravida nunc id ipsum vestibulum rhoncus. Quisque ultrices turpis et purus dapibus sollicitudin. Integer quis scelerisque sapien, et blandit ex. Proin lacinia sem vitae neque facilisis scelerisque. Mauris gravida ex et justo hendrerit suscipit. Fusce vel leo mi. Mauris vitae elementum sapien. Donec aliquet tempus quam ac pulvinar. Vestibulum ultricies odio id aliquam laoreet. Quisque nec posuere lorem. Quisque eget nisi pellentesque, cursus felis hendrerit, venenatis ante. Phasellus sed risus vitae lorem viverra vulputate. Phasellus fermentum odio vel mollis auctor.
+{% CustomFont fontSettings={fontFamily: {discriminant: false}, fontWeight: {discriminant: false}, fontStyle: {discriminant: true, value: "italic"}, fontSize: {discriminant: false}, lineHeight: {discriminant: false}, textColor: {discriminant: false}} %}Mauris gravida nunc id ipsum vestibulum rhoncus. Quisque ultrices turpis et purus dapibus sollicitudin. Integer quis scelerisque sapien, et blandit ex. Proin lacinia sem vitae neque facilisis scelerisque. Mauris gravida ex et justo hendrerit suscipit. Fusce vel leo mi. Mauris vitae elementum sapien. Donec aliquet tempus quam ac pulvinar. Vesti{% /CustomFont %}bulum ultricies odio id aliquam laoreet. Quisque nec posuere lorem. Quisque eget nisi pellentesque, cursus felis hendrerit, venenatis ante. Phasellus sed risus vitae lorem viverra vulputate. Phasellus fermentum odio vel mollis auctor.
 
 Ut sit amet semper est. Etiam tincidunt, leo nec auctor porta, magna lacus placerat felis, ut ultricies tellus nunc sit amet nulla. Pellentesque dignissim viverra eros ac accumsan. Integer orci diam, tincidunt et sapien eu, gravida accumsan mauris. Maecenas et fermentum orci. Nullam nec mi eget tortor ullamcorper semper. Fusce finibus purus nec neque varius varius.
